Capture details for "About you" step
In its simplest form, and with no validation, integrate Wicked::Wizard.
Use the MembershipApplicationsController to create applications only
with a first name and email. Use a session variable to hold the ID of
the current membership application.

Then use the MembershipApplications::StepsController to build up the
MembershipApplication object. Currently, that means capturing the
details, not validating them at all, and then redirecting to a dummy
"completed" page which holds no content at all.