Add contact-details step
This is the first step that requires a fleshed-out version of a
MembershipApplication::Steps class that can tell us if we reached a
step.

To do this we define a Singleton Steps::PARAMS hash that has the step
names on the left-hand side and the params names on the right. It works
out the indices for each and allows us to ask Steps.instance
.reached_step? for any given step.

While we're about it, we can, at any step, only permit the parameters
expected at that step in StepsController.