Add a help_for(field) helper
This renders the little bit of grey text along with a link for a
dropdown if necessary. All content comes from en.yml. Add a help_text
key for just text, and add a help_dropdown key with two subkeys `link`
and `content` for optional dropdown support.

The helper is just sugar for rendering shared/help_for_field, which is
a lot easier to maintain than trying to use content_tag in helpers.