Make a date range search
Available from both HTML and JSON. Use a view model
View::DateRangeQuery to validate the parameters, and render errors
as JSON when necessary.

It's a bit hacky on the form side, using a combination of

form_with scope: nil and an `f.object = @date_range_query` to get
our field-with-errors implementation working.

Also we're shoehorning some help text for the search in with
membership_application in en.yml. It doesn't really belong there but
we don't have to refactor help_for and its template to get it.