Avoid initializer autoload deprecation
```
DEPRECATION WARNING: Initialization autoloaded the constants ApplicationRecord, MembershipApplication, and MembershipApplication::Steps.

Being able to do this is deprecated. Autoloading during initialization is going to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during initialization does not run again. So, if you reload ApplicationRecord, for example, the expected changes won't be reflected in that stale Class object.

These autoloaded constants have been unloaded.
 ```